Consolidating MLN and CP to DL4J
