+++ konduit-serving-api/src/main/java/ai/konduit/serving/model/ModelConfig.java
@@ -88,8 +88,7 @@ public class ModelConfig implements Serializable {
 
 
     public enum ModelType {
-        COMPUTATION_GRAPH,
-        MULTI_LAYER_NETWORK,
+        DL4J,
         PMML,
         TENSORFLOW,
         KERAS,
