+++ konduit-serving-test/src/test/java/ai/konduit/serving/configprovider/KonduitServingMainTest.java
@@ -147,7 +147,7 @@ public class KonduitServingMainTest {
         ModelConfig modelConfig = ModelConfig.builder()
                 .modelConfigType(
                         ModelConfigType.builder().modelLoadingPath(modelSave.getAbsolutePath())
-                                .modelType(ModelConfig.ModelType.MULTI_LAYER_NETWORK)
+                                .modelType(ModelConfig.ModelType.DL4J)
                                 .build()
                 ).build();
