+++ konduit-serving-test/src/test/java/ai/konduit/serving/verticles/ndarray/BaseDl4JVerticalTest.java
@@ -97,7 +97,7 @@ public abstract class BaseDl4JVerticalTest extends BaseVerticleTest {
         Nd4j.getRandom().setSeed(42);
 
         ModelConfig modelConfig = ModelConfig.builder()
-                .modelConfigType(ModelConfigType.multiLayerNetwork(modelSave.getAbsolutePath()))
+                .modelConfigType(ModelConfigType.dl4j(modelSave.getAbsolutePath()))
                 .build();
 
         ServingConfig servingConfig = ServingConfig.builder()
