+++ konduit-serving-test/src/test/java/ai/konduit/serving/verticles/ndarray/ColumnarTransformProcessesTest.java
@@ -118,7 +118,7 @@ public class ColumnarTransformProcessesTest extends BaseDl4JVerticalTest {
                 .build();
 
         ModelConfig modelConfig = ModelConfig.builder()
-                .modelConfigType(ModelConfigType.multiLayerNetwork(modelSave.getAbsolutePath()))
+                .modelConfigType(ModelConfigType.dl4j(modelSave.getAbsolutePath()))
                 .build();
 
         PipelineStep modelStepConfig = new ModelStep(modelConfig)
